backend/requirements.txt
torch==2.0.1
pandas
lxml
crewai[tools]
pyautogen[retrievechat]
unstructured[all-docs]
langchain 
langchain-cli
openai 
weaviate-client
pinecone-client
python-dotenv
streamlit
langchain-community
fastapi[all]
uvicorn
bitsandbytes
langchain-community
transformers
accelerate
faiss-gpu
jq
python-multipart
pymongo
tenacity==8.3.0